+++ bindings/nostr-sdk-js/src/relay/limits.rs
@@ -52,10 +52,30 @@ impl JsRelayLimits {
         self
     }
 
+    /// Maximum size per kind of normalised JSON, in bytes
+    #[wasm_bindgen(js_name = eventMaxSizePerKind)]
+    pub fn event_max_size_per_kind(mut self, kind: u16, max_size: Option<u32>) -> Self {
+        self.inner.events = self
+            .inner
+            .events
+            .set_max_size_per_kind(Kind::from(kind), max_size);
+        self
+    }
+
     /// Maximum number of tags allowed (default: 2_000)
     #[wasm_bindgen(js_name = eventMaxNumTags)]
     pub fn event_max_num_tags(mut self, max_num_tags: Option<u16>) -> Self {
         self.inner.events.max_num_tags = max_num_tags;
+        self
+    }
+
+    /// Maximum number of tags per kind allowed
+    #[wasm_bindgen(js_name = eventMaxNumTagsPerKind)]
+    pub fn event_max_num_tags_per_kind(mut self, kind: u16, max_num_tags: Option<u16>) -> Self {
+        self.inner.events = self
+            .inner
+            .events
+            .set_max_num_tags_per_kind(Kind::from(kind), max_num_tags);
         self
     }
 }